complete backend

+++ src/index.ts
@@ -2,12 +2,14 @@ import { Express, Request, Response, urlencoded } from "express";
 
 const express = require("express");
 const app: Express = express();
+const cors = require("cors");
 
+app.use(cors());
 app.use(urlencoded({ extended: true }));
 app.use(express.json());
 
 const jwt = require("jsonwebtoken");
-const auth = require("./utils/auth");
+const { auth } = require("./utils/auth");
 
 const { connectDB } = require("./utils/database");
 const { ItemModel, UserModel } = require("./utils/schemaModels");
@@ -57,10 +59,16 @@ app.get("/item/:id", async (req: Request, res: Response) => {
 app.put("/item/update/:id", auth, async (req: Request, res: Response) => {
   try {
     await connectDB();
-    await ItemModel.updateOne({ _id: req.params.id }, req.body);
-    return res.status(200).json({
-      message: "アイテム編集成功",
-    });
+    const singleItem = await ItemModel.findById(req.params.id);
+
+    if (singleItem.email === req.body.email) {
+      await ItemModel.updateOne({ _id: req.params.id }, req.body);
+      return res.status(200).json({
+        message: "アイテム編集成功",
+      });
+    } else {
+      throw new Error();
+    }
   } catch (err) {
     return res.status(400).json({
       message: "アイテム編集失敗",
@@ -71,8 +79,14 @@ app.put("/item/update/:id", auth, async (req: Request, res: Response) => {
 app.delete("/item/delete/:id", auth, async (req: Request, res: Response) => {
   try {
     await connectDB();
-    await ItemModel.deleteOne({ _id: req.params.id });
-    return res.status(200).json({ message: "アイテム削除成功" });
+    const singleItem = await ItemModel.findById(req.params.id);
+
+    if (singleItem.email === req.body.email) {
+      await ItemModel.deleteOne({ _id: req.params.id });
+      return res.status(200).json({ message: "アイテム削除成功" });
+    } else {
+      throw new Error();
+    }
   } catch (err) {
     return res.status(400).json({
       message: "アイテム削除失敗",
@@ -134,4 +148,11 @@ app.post("/user/login", async (req: Request, res: Response) => {
 
 app.listen(5000, () => {
   console.log("Listening on localhost port 5000");
+});
+
+// Connecting to port
+const port = process.env.PORT || 5000;
+
+app.listen(port, () => {
+  console.log(`Listening on localhost port ${port}`);
 });
